controller y Dao

+++ src/main/java/pe/edu/cibertec/RC/Repository/ProductDao.java
@@ -3,6 +3,7 @@ package pe.edu.cibertec.RC.Repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import pe.edu.cibertec.RC.Model.Productos;
 
-public interface ProductDao extends JpaRepository<Productos, Long> {
+public interface ProductDao extends JpaRepository<Productos, Integer> {
+
 
 }
